+++ CHANGELOG.rst
@@ -5,6 +5,7 @@
 - Respect `temporary-directory` config for spilling (#247) `John Kirkham`_
 - Relax CuPy pin (#248) `John Kirkham`_
 - Added `ignore_index` argument to `partition_by_hash()` (#253) `Mads R. B. Kristensen`_
+- Drop Numba `DeviceNDArray` code for `sizeof` (#257) `John Kirkham`_
 
 0.12
 ----